[FIX] base@0.0.0: do not delete studio views when they fail
The studio module creates automatically a fake module in
ir.module.module:

https://github.com/odoo/enterprise/blob/4321348535082741f99c534dd0897b4382e82d0f/web_studio/models/ir_module_module.py#L18-L29

However, the views created by studio should not be deleted,
because they won't be restored automatically when doing a -u on the module,
as there is no module studio_customization : D.

In addition, I add the check on view.arch_fs,
because if the view xml_id tells it comes from a custom module,
but there is no associated file (stored in view.arch_fs),
then this is possible this is not actually a view coming from the given
module, as there is no file associated to it. In such a case,
the view should not be deleted either

+++ migrations/base/0.0.0/pre-models-ir_ui_view.py
@@ -50,7 +50,7 @@ class IrUiView(models.Model):
             # so if there are multiple custom views but only one causes an issue, it tries to disable only this one.
             standard_modules = get_modules()
             custom_modules = [r['name'] for r in self.env['ir.module.module'].search_read([
-                ('name', 'not in', standard_modules)
+                ('name', 'not in', standard_modules + ['studio_customization'])
             ], ['name'])]
             views_to_check = []
             views = self._migrate_get_roots()
@@ -84,7 +84,7 @@ class IrUiView(models.Model):
                     The custom view `%s` (ID: %s, Inherit: %s, Model: %s) caused validation issues.
                     Disabling it for the migration ...
                     """, view.name, view.id, view.inherit_id, view.model)
-                    if md.module in custom_modules:
+                    if md.module in custom_modules and view.arch_fs:
                         # custom view in a custom module, to remove
                         # it will be re-created when the user call -u on its custom module
                         view.unlink()
